% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooperatives.R
\docType{data}
\name{cooperatives}
\alias{cooperatives}
\title{Characteristics of 67 fishing cooperatives around the world}
\format{A data frame with 103 rows and 133 variables}
\usage{
data(cooperatives)
}
\description{
This document provides definitions for variables
collected in the database of cooperative fisheries. Data are
provided in broad categories of


\itemize{
  \item General information
  \item Ecosystem characteristics
  \item Host country
  \item Catch share management
  \item Non-catch share management
  \item Whole fleet data
  \item Fishery specific data
  \item Cooperative behaviors
  \item Cooperative characteristics
  \item Exclusivity
  \item Role of government
}

Variables fall into 2 general classes:

\itemize{
  \item binary data (1 or 0)
  \item descriptive entry (e.g. 10,000 tones, coral reef)
}

All variable names are written in snake case. For example,
a variable for consistent spawining location of a species
is included as \code{consistent_spawning_location}.

A full list of variables present in the dataset can be seen
by using the \code{\link{get_variables}} function.

For a full documentation of the original dataset, look at the supplementary
information of the paper in: https://doi.org/10.1016/j.marpol.2012.03.012.

For a record of changes made to the database visit http://www.jcvdav.github.io/cooperatR/DBcleaning.nb.html
}
\references{
Daniel A. Ovando, Robert T. Deacon, Sarah E. Lester, Christopher Costello, Tonya Van Leuvan, Karlynn McIlwain, C. Kent Strauss, Michael Arbuckle, Rod Fujita, Stefan Gelcich, Hirotsugu Uchida, Conservation incentives and collective choices in cooperative fisheries, Marine Policy, Volume 37, January 2013, Pages 132-140, ISSN 0308-597X, https://doi.org/10.1016/j.marpol.2012.03.012.
}
\seealso{
\code{\link{get_variables}}
}
\author{
Juan Carlos
}
\keyword{datasets}
