% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp.R
\name{get_spp}
\alias{get_spp}
\title{Get species names}
\usage{
get_spp(scientific = FALSE)
}
\arguments{
\item{scientific}{If \code{scientific = TRUE}, then a list of scientific names is returned. If \code{scientific = FALSE} (the default) only common names are returned.}
}
\value{
A character vector containing a unique list of species present in the cooperatives dataset.
}
\description{
Looks at the database and extracts a list of all species present. The user can specify if common or scientific names should be returned (See scientific).
}
\examples{

#Get a list of the 86 distinct common names
get_spp(scientific = FALSE)


#Get a list of the 83 distinct scientific names
get_spp(scientific = TRUE)

}
\seealso{
\code{\link{cooperatives}}
}
\author{
Juan Carlos
}
