% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique.R
\name{get_unique}
\alias{get_unique}
\title{Get unique values}
\usage{
get_unique(other.variables = NULL, include.spp = TRUE, to.file = FALSE,
  filename = NULL)
}
\arguments{
\item{other.variables}{A character string specifying if other
variables should be included. See the \code{\link{get_variables}}
function for a list of variables available in \code{\link{cooperatives}}.
If not specified, it defaults to \code{NULL} and no additional
variables are returned.}

\item{include.spp}{A logical indicating if the columns for
\code{target_spp} and \code{complete_name} should be included.
By default, it is set to \code{TRUE}.}

\item{to.file}{A logical indicating if the extraction should be
written to a csv file on disk.}

\item{filename}{If the \code{to.file} argument is set to
\code{TRUE}, the user may want to specify a particular filename.
This must be a character string that specifies the desired name
of the file. The user can also specify the directory relative
to its current working directory with a customary approach of
"./RootFolder/MyDesiredFolder/filename". There is no need to
specify file extension! If \code{NULL} (de default),
the file is saved as cooperatives.csv in the current working directory.}
}
\value{
a \code{data.frame} containing the desired columns
}
\description{
Creates a \code{data.frame} containing columns
that uniquely identify a fishery in the \code{\link{cooperatives}}
dataset. By default, it includes \code{original_order},
\code{entered_by}, \code{fishery_id},
\code{geographic_range_of_fishing_acti}, and \code{home_port}.
}
\examples{

# Get unique columns excluding species and save to file
get_unique(include.spp = FALSE, to.file = TRUE)

# Get unique columns, with species, and include \\code{support_of_voluntary_regs} and \\code{gear_type} columns
get_unique(other.variables = c("support_of_voluntary_regs", "gear_type"))

}
\seealso{
\code{\link{cooperatives}}, \code{\link{get_variables}}
}
\author{
Juan Carlos
}
