% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Get variable names of the \code{\link{cooperatives}} dataset}
\usage{
get_variables(sort = FALSE)
}
\arguments{
\item{sort}{if \code{sort = FALSE} (the default) variable names are returned in the original order of the database (usefull, because they are grouped by type). If \code{sort = TRUE} variable names are sorted alphabetically (usefull to inspect existing variables).}
}
\value{
a character vector containing the names of the variables in the dataset
}
\description{
Extracts variable names from the \code{\link{cooperatives}} dataset
}
\examples{

# Extract variable names in original order
get_variables(sort = FALSE)

# Extract sorted variable names
get_variables(sort = TRUE)

}
\seealso{
\code{\link{cooperatives}}
}
