% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save2csv.R
\name{save2csv}
\alias{save2csv}
\title{Save the \code{\link{cooperatives}} dataset to a *.csv file}
\usage{
save2csv(filename = NULL)
}
\arguments{
\item{filename}{A character string that specifies the desired name of the file. The user can also specify the directory relative to its current working directory with a customary approach of "./RootFolder/MyDesiredFolder/filename". There is no need to specify file extension! If \code{NULL} (de default), the file is saved as cooperatives.csv in the current working directory.}
}
\description{
Saves the entire cooperatives dataset to a *.csv file that can be manipulated by the user. BUT any modifications done to the file are not replicable and difficult to incorporate into the original database!
}
\examples{

# Save database with default name
save2csv()

# Save to current directory with custom name
save2csv(filename = "coop_data")

# Save database to a desired working directory
save2csv(filename = "./Data/coop_data")

}
\seealso{
\code{\link{cooperatives}}
}
\author{
Juan Carlos
}
